Update Submit_UserSubmission.py
Now jobs will not get stuck in queue if they fail and stop up the whole line. As configured, jobs will try to be submitted once, and then be marked as failures and not tried again. If instead you want failed job submissions to be resubmitted, change the value of submission_string from "Submitted to Failure Mode" to e.g. "Failed Submission". The codebase (database.get_unsubmitted_jobs()) will try to submit any jobs with run_status field NOT LIKE "Submitted to*".